pyproject.toml
[project]
name = "estate-and-tax-document-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.0",
    "openai>=1.98.0",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "tesseract>=0.1.3",
]
